+++ Content.Server/Chemistry/Components/ConvectionHeatComponent.cs
@@ -1,0 +1,8 @@
+namespace Content.Server.Chemistry.Components;
+
+[RegisterComponent]
+public sealed partial class ConvectionHeatComponent : Component
+{
+    [DataField, ViewVariables(VVAccess.ReadWrite)]
+    public float TempDifference;
+}
